+++ src/Generics/Deriving/Base.hs
@@ -8,6 +8,10 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE TypeSynonymInstances #-}
 
+#if __GLASGOW_HASKELL__ >= 701
+{-# LANGUAGE Trustworthy #-}
+#endif
+
 module Generics.Deriving.Base (
 -- * Introduction
 --